Add modal loading dialog
This dialog will prevent the user from interacting with the program
while pywebview initializes itself in the background. Previously,
the program would wait for a static amount of time and assume that
pywebview would be loaded by the end of the delay period. There is
now an appropriate check in place that will keep the dialog visible
and the program waiting until pywebview is known to be initialized.